[FIX] mrp: Share Google Slide viewer
As the previous 'embed_viewer' was also (wrongly) used in maintenance,
make it available globally.
