+++ addons/mrp/static/src/views/fields/google_slides_viewer.xml
@@ -1,17 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-    <t t-name="mrp.SlidesViewer" t-inherit="web.CharField">
-        <xpath expr="//t[@t-else='']" position="after">
-            <t t-if="url">
-                <iframe class="o_embed_iframe w-100"
-                    alt="Slides viewer"
-                    t-att-src="url"
-                    t-att-name="props.name"
-                    t-on-error="onLoadFailed"
-                    />
-            </t>
-            </xpath>
-        </t>
+    <t t-name="mrp.SlidesViewer" t-inherit="web.GoogleSlideViewer" />
 
 </templates>
