+++ addons/web/static/src/views/fields/google_slide_viewer/google_slide_viewer.scss
@@ -1,0 +1,17 @@
+.o_field_widget.o_field_google_slide_viewer {
+    width: 100%;
+}
+.o_field_widget.o_field_google_slide_viewer div {
+    position: relative;
+    padding-bottom: 56.25%; /* 16:9 Ratio */
+    height: 0;
+    overflow: hidden;
+}
+.o_field_widget.o_field_google_slide_viewer iframe {
+    border: 0;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100% !important;
+    height: 100% !important;
+}
