+++ addons/web/static/src/views/fields/google_slide_viewer/google_slide_viewer.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+
+    <t t-name="web.GoogleSlideViewer" t-inherit="web.CharField">
+        <xpath expr="//t[@t-else='']" position="after">
+            <t t-if="url">
+                <div>
+                    <iframe class="o_embed_iframe w-100"
+                        alt="Google Slide Viewer"
+                        t-att-src="url"
+                        t-att-name="props.name"
+                        t-on-error="onLoadFailed"
+                    />
+                </div>
+            </t>
+        </xpath>
+    </t>
+
+</templates>
